+++ base/src/jvmMain/kotlin/utils/Decimal.kt
@@ -91,14 +91,14 @@ actual fun Decimal.round(scale: Int, roundingMode: RoundingMode) =
         )
     )
 
-actual infix fun Decimal.pow(n: UInt): Decimal = this.pow(n.toInt(), MathContext.DECIMAL128)
-actual fun Decimal.pow(n: UInt, scale: Int): Decimal = this.pow(n.toInt(), MathContext.DECIMAL128).setScale(scale, NativeRoundingMode.HALF_EVEN)
+actual infix fun Decimal.pow(n: Int): Decimal = this.pow(n, MathContext.DECIMAL128)
+actual fun Decimal.pow(n: Int, scale: Int): Decimal = this.pow(n, MathContext.DECIMAL128).setScale(scale, NativeRoundingMode.HALF_EVEN)
 actual fun Decimal.pow(
-    n: UInt,
+    n: Int,
     scale: Int,
     roundingMode: RoundingMode
 ): Decimal = this.pow(
-    n.toInt(),
+    n,
     MathContext(
         MathContext.DECIMAL128.precision,
         NativeRoundingMode.valueOf(roundingMode.java)
